fix supabase
